# docs/baseline/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DataResponse_PolicyVersionsResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PolicyVersionsResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaResponse",
            "default": {}
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[PolicyVersionsResponse]",
        "type": "object"
      },
...
      "DataResponse_dict_str__Any__": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaResponse",
            "default": {}
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[dict[str, Any]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MetaResponse": {
        "properties": {
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "title": "MetaResponse",
        "type": "object"
      },
...
      "PolicyCreateRequest": {
        "properties": {
          "allow": {
            "additionalProperties": true,
            "description": "Allow block: caller_type, callee_type, capability_types",
            "title": "Allow",
            "type": "object"
          },
          "conditions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Conditions",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "hil_threshold_usd": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hil Threshold Usd"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "on_violation": {
            "default": "block_and_alert",
            "title": "On Violation",
            "type": "string"
          },
          "priority": {
            "default": 100,
            "description": "Lower = evaluated first",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "allow"
        ],
        "title": "PolicyCreateRequest",
        "type": "object"
      },
      "PolicyResponse": {
        "properties": {
          "allow": {
            "additionalProperties": true,
            "title": "Allow",
            "type": "object"
          },
          "conditions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Conditions",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "hil_threshold_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hil Threshold Usd"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "on_violation": {
            "title": "On Violation",
            "type": "string"
          },
          "parent_policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Policy Id"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "parent_policy_id",
          "name",
          "description",
          "priority",
          "version",
          "enabled",
          "allow",
          "conditions",
          "hil_threshold_usd",
          "on_violation",
          "created_at"
        ],
        "title": "PolicyResponse",
        "type": "object"
      },
      "PolicyUpdateRequest": {
        "properties": {
          "allow": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow"
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "hil_threshold_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hil Threshold Usd"
          },
          "on_violation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Violation"
          },
          "priority": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "title": "PolicyUpdateRequest",
        "type": "object"
      },
      "PolicyVersionsResponse": {
        "properties": {
          "policy_id": {
            "title": "Policy Id",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/PolicyResponse"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "policy_id",
          "versions"
        ],
        "title": "PolicyVersionsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The control plane for AI agent networks",
    "title": "Nexra API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/policies": {
      "get": {
        "description": "List all policies for the authenticated org.",
        "operationId": "list_policies_v1_policies_get",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_str__Any__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Policies",
        "tags": [
          "policies"
        ]
      },
      "post": {
        "description": "Create a new delegation policy.",
        "operationId": "create_policy_v1_policies_post",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Policy",
        "tags": [
          "policies"
        ]
      }
    },
    "/v1/policies/{policy_id}": {
      "delete": {
        "description": "Disable a policy (soft delete). Preserves history.",
        "operationId": "disable_policy_v1_policies__policy_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Policy",
        "tags": [
          "policies"
        ]
      },
      "get": {
        "description": "Get a policy by ID with version history.",
        "operationId": "get_policy_v1_policies__policy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_str__Any__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Policy",
        "tags": [
          "policies"
        ]
      },
      "put": {
        "description": "Update a policy. Creates a new version; old version is preserved.",
        "operationId": "update_policy_v1_policies__policy_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Policy",
        "tags": [
          "policies"
        ]
      }
    },
    "/v1/policies/{policy_id}/versions": {
      "get": {
        "description": "Get immutable version history for a policy family.",
        "operationId": "get_policy_versions_v1_policies__policy_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PolicyVersionsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Policy Versions",
        "tags": [
          "policies"
        ]
      }
    },
...
  }
}